add lenses to cart

+++ database/migrations/2023_05_19_203035_create_cart_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained('products');
             $table->foreignId('cart_id')->constrained('carts');
-            $table->double('quantity');
+            $table->integer('quantity');
             $table->timestamps();
         });
     }
